NegraCorpusReader uses the same format as other parsed corpus readers
NEGRA and TIGER are parsed corpora, not chunked corpora.  Rename
chunked_sents to parsed_sents.  The Trees returned by this method are
in the same format as is returned by other parsed corpora in NLTK.

Specifically, the words in parse trees that were previously leaf nodes
represented as (word, tag) are now Tree(tag, [word]).